valyuta kursi/requirements.txt
python-telegram-bot==13.13
requests
